Added Contoller Endpoints for task three microservice, updated firestoreaccess in task three micro service to allow fund transfer via IBAN instead of account no when transfering to third party accounts

+++ Task1MicroService/MicroServiceThree/Model/FundAccount.cs
@@ -28,10 +28,10 @@ namespace MicroServiceThree.Model
         public string CurrencyCode { get; set; }
 
         [FirestoreProperty]
-        public float OpeningBalance { get; set; }
+        public double OpeningBalance { get; set; }
 
         [FirestoreProperty]
-        public float Balance { get; set; }
+        public double Balance { get; set; }
 
         [FirestoreProperty]
         public string Status { get; set; } // to check if status is disable or not-disabled
